requirements.txt
fastapi==0.111.0
python-dotenv==1.0.1
pydantic==2.7.2
pydantic-settings==2.2.1
requests==2.32.3
openai==1.30.2
#pillow==9.5.0
python-docx==1.1.2
PyMuPDF==1.24.5
requests==2.32.3
aiohttp==3.9.5
tenacity==8.5.0

numpy==1.21.4  # Ensure NumPy 1.x version is installed
uvicorn
anthropic
pytesseract
pdfplumber
pytest
layoutparser


